added error handling, form validation + minor bugfixes

+++ routes/web.php
@@ -22,12 +22,12 @@ Route::get('/dashboard', function () {
 })->middleware(['auth'])->name('dashboard');
 
 Route::resource('products','App\Http\Controllers\ProductsController');
-Route::get('products',['App\Http\Controllers\ProductsController','index'])->name('products.index');
-Route::get('products/create',['App\Http\Controllers\ProductsController','create'])->name('products.create');
-Route::get('pricelist',['App\Http\Controllers\ProductsController','pricelist'])->name('products.pricelist');
-Route::post('products/create',['App\Http\Controllers\ProductsController','store'])->name('products.store');
-Route::post('products/pricelist',['App\Http\Controllers\ProductsController','storeprice'])->name('products.storeprice');
-Route::post('products',['App\Http\Controllers\ProductsController','filter'])->name('products.filter');
-Route::delete('pricelist/{id}',['App\Http\Controllers\ProductsController','destroyprice'])->name('products.destroyprice');
+Route::get('products',['App\Http\Controllers\ProductsController','index'])->middleware(['auth'])->name('products.index');
+Route::get('products/create',['App\Http\Controllers\ProductsController','create'])->middleware(['auth'])->name('products.create');
+Route::get('pricelist',['App\Http\Controllers\ProductsController','pricelist'])->middleware(['auth'])->name('products.pricelist');
+Route::post('products/create',['App\Http\Controllers\ProductsController','store'])->middleware(['auth'])->name('products.store');
+Route::post('products/pricelist',['App\Http\Controllers\ProductsController','storeprice'])->middleware(['auth'])->name('products.storeprice');
+Route::post('products',['App\Http\Controllers\ProductsController','filter'])->middleware(['auth'])->name('products.filter');
+Route::delete('pricelist/{id}',['App\Http\Controllers\ProductsController','destroyprice'])->middleware(['auth'])->name('products.destroyprice');
 
 require __DIR__.'/auth.php';
